+++ lib/opal.rb
@@ -6,6 +6,7 @@ require 'opal/paths'
 require 'opal/version'
 require 'opal/errors'
 require 'opal/source_map'
+require 'opal/deprecations'
 
 # Opal is a ruby to javascript compiler, with a runtime for running
 # in any JavaScript environment.