+++ lib/opal/server.rb
@@ -1,3 +1,3 @@
-warn "DEPRECATED: `require 'opal/server` and `Opal::Server` are deprecated in favor of `require 'opal/sprockets/server'` and `Opal::Sprockets::Server` (now part of the opal-sprockets gem)."
+Opal.deprecation "`require 'opal/server` and `Opal::Server` are deprecated in favor of `require 'opal/sprockets/server'` and `Opal::Sprockets::Server` (now part of the opal-sprockets gem)."
 require 'opal/sprockets/server'
 Opal::Server = Opal::Sprockets::Server